extract libjpeg logic for android build
