+++ rimage/image.go
@@ -1,5 +1,3 @@
-//go:build !no_cgo
-
 // Package rimage defines fundamental image and color processing primitives.
 //
 // The golang standard library, while useful, is not very productive when it
